pipa: libc: Force build scudo free 32 bit variant of libc.
We have issues with camera blobs which doesn't play
well with scudo allocation/deallocation tracking.
Scudo reports corrupted header when stopping a
video recording.

Because we aren't alone with this (also upstream seems
to have trouble with camera modules) use the 32 libc
which is used only by a few components, like camera
and audio.

Because docs states, that libc with the svelte take less RSS,
but is a little bit slower.
https://android.googlesource.com/platform/bionic/+/master/docs/native_allocator.md
This is the reason to not enable this target wide.

Change-Id: Ia4d3da9d20fa4c5633496efdd210ad812d9e0280